ForgeTM/apps/backend/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "forge-backend"
version = "0.1.0"
description = "ForgeTM Backend API"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.114.2",
  "uvicorn[standard]==0.30.6",
  "httpx==0.27.2",
  "python-dotenv==1.0.1",
  "pydantic==2.9.2",
  "pydantic-settings==2.3.0",
  "mypy==1.11.2",
  "opentelemetry-distro==0.48b0",
  "opentelemetry-exporter-otlp==1.27.0",
  "opentelemetry-instrumentation-fastapi==0.48b0",
  "opentelemetry-instrumentation-httpx==0.48b0",
  "opentelemetry-sdk==1.27.0",
  "sentry-sdk[fastapi]==2.13.0",
  "celery[redis]==5.4.0",
  "redis==5.0.8",
  "openfeature-sdk==0.7.0",
  "litellm==1.48.0",
  "pre-commit==3.8.0",
  "hypothesis==6.115.2",
  "coverage==7.6.1",
  "alembic==1.13.3",
  "sqlalchemy==2.0.35",
  "pinecone-client==4.1.0",
  "python-jose[cryptography]==3.3.0",
  "passlib[bcrypt]==1.7.4",
  "requests==2.32.3",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
forge = "src/forge"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = []

[tool.ruff.format]
quote-style = "single"

[project.optional-dependencies]
dev = [
  "ruff==0.6.9",
  "pytest==8.3.2",
  "pytest-cov==5.0.0",
  "pytest-asyncio==0.23.8",
  "hypothesis==6.115.2",
  "mkdocs==1.5.3",
  "mkdocs-material==9.5.18",
  "mkdocstrings[python]==0.25.1",
  "nox==2024.4.15",
  "pre-commit==3.8.0",
  "coverage==7.6.1",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[dependency-groups]
dev = [
  "ruff==0.6.9",
  "pytest==8.3.2",
  "pytest-cov==5.0.0",
  "pytest-asyncio==0.23.8",
  "hypothesis==6.115.2",
  "mkdocs==1.5.3",
  "mkdocs-material==9.5.18",
  "mkdocstrings[python]==0.25.1",
  "nox==2024.4.15",
  "pre-commit==3.8.0",
  "coverage==7.6.1",
]
